pyproject.toml
[tool.poetry]
name = "sd-jwt"
version = "0.9.0"
description = "The reference implementation of the IETF SD-JWT specification."
authors = ["Daniel Fett <mail@danielfett.de>"]
readme = "README.md"
packages = [{include = "sd_jwt"}]

[tool.poetry.dependencies]
python = "^3.8"
jwcrypto = ">=1.3.1"
pyyaml = ">=5.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sd-jwt = "sd_jwt.bin.sd_jwt:main"
